move TUS related test-steps
