+++ tests/acceptance/features/bootstrap/WebDav.php
@@ -34,7 +34,6 @@ use TestHelpers\HttpRequestHelper;
 use TestHelpers\Asserts\WebDav as WebDavAssert;
 use TusPhp\Exception\ConnectionException;
 use TusPhp\Exception\TusException;
-use TusPhp\Tus\Client;
 
 /**
  * WebDav functions
@@ -1600,59 +1599,6 @@ trait WebDav {
 	}
 
 	/**
-	 * @When user :user uploads file :source to :destination using the TUS protocol on the WebDAV API
-	 *
-	 * @param string $user
-	 * @param string $source
-	 * @param string $destination
-	 * @param array  $uploadMetadata array of metadata to be placed in the
-	 *                               `Upload-Metadata` header.
-	 *                               see https://tus.io/protocols/resumable-upload.html#upload-metadata
-	 *                               Don't Base64 encode the value.
-	 * @param int    $noOfChunks
-	 *
-	 * @return void
-	 * @throws ConnectionException
-	 * @throws ReflectionException
-	 * @throws TusException
-	 */
-	public function userUploadsUsingTusAFileTo(
-		string $user,
-		string $source,
-		string $destination,
-		array $uploadMetadata = [],
-		int $noOfChunks = 1
-	) {
-		$user = $this->getActualUsername($user);
-		$password = $this->getUserPassword($user);
-		$client = new Client(
-			$this->getBaseUrl(),
-			['verify' => false,
-				'headers' => [
-				'Authorization' => 'Basic ' . \base64_encode($user . ':' . $password)
-				]
-			]
-		);
-		$client->setApiPath(
-			WebDavHelper::getDavPath($user, $this->getDavPathVersion())
-		);
-		$client->setMetadata($uploadMetadata);
-		$sourceFile = $this->acceptanceTestsDirLocation() . $source;
-		$client->setKey(\rand())->file($sourceFile, $destination);
-		$this->pauseUploadDelete();
-
-		if ($noOfChunks === 1) {
-			$client->file($sourceFile, $destination)->upload();
-		} else {
-			$bytesPerChunk = \ceil(\filesize($sourceFile) / $noOfChunks);
-			for ($i = 0; $i < $noOfChunks; $i++) {
-				$client->upload($bytesPerChunk);
-			}
-		}
-		$this->lastUploadDeleteTime = \time();
-	}
-
-	/**
 	 * @Given user :user has uploaded file :source to :destination
 	 *
 	 * @param string $user
@@ -2282,72 +2228,6 @@ trait WebDav {
 	}
 
 	/**
-	 * @param string $content
-	 *
-	 * @return string the file name
-	 * @throws Exception
-	 */
-	private function writeDataToTempFile(string $content) {
-		$tmpfname = \tempnam(
-			$this->acceptanceTestsDirLocation(), "tus-upload-test-"
-		);
-		if ($tmpfname === false) {
-			throw new \Exception("could not create a temporary filename");
-		}
-		$tempfile = \fopen($tmpfname, "w");
-		if ($tempfile === false) {
-			throw new \Exception("could not open " . $tmpfname . " for write");
-		}
-		\fwrite($tempfile, $content);
-		\fclose($tempfile);
-		return $tmpfname;
-	}
-	/**
-	 * @When user :user uploads file with content :content to :destination using the TUS protocol on the WebDAV API
-	 *
-	 * @param string $user
-	 * @param string $content
-	 * @param string $destination
-	 *
-	 * @return string
-	 */
-	public function userUploadsAFileWithContentToUsingTus(
-		string $user, string $content, string $destination
-	) {
-		$tmpfname = $this->writeDataToTempFile($content);
-		$this->userUploadsUsingTusAFileTo(
-			$user, \basename($tmpfname), $destination
-		);
-		\unlink($tmpfname);
-	}
-
-	/**
-	 * @When user :user uploads file with content :content in :noOfChunks chunks to :destination using the TUS protocol on the WebDAV API
-	 *
-	 * @param string $user
-	 * @param string $content
-	 * @param int    $noOfChunks
-	 * @param string $destination
-	 *
-	 * @return void
-	 * @throws ConnectionException
-	 * @throws ReflectionException
-	 * @throws TusException
-	 */
-	public function userUploadsAFileWithContentInChunksUsingTus(
-		string $user,
-		string $content,
-		int $noOfChunks,
-		string $destination
-	) {
-		$tmpfname = $this->writeDataToTempFile($content);
-		$this->userUploadsUsingTusAFileTo(
-			$user, \basename($tmpfname), $destination, [], $noOfChunks
-		);
-		\unlink($tmpfname);
-	}
-
-	/**
 	 * @When user :user uploads file :source to :destination with mtime :mtime using the WebDAV API
 	 * @Given user :user has uploaded file :source to :destination with mtime :mtime using the WebDAV API
 	 *
@@ -2368,27 +2248,6 @@ trait WebDav {
 			$this->getBaseUrl(), $user, $this->getPasswordForUser($user),
 			$this->acceptanceTestsDirLocation() . $source, $destination,
 			["X-OC-Mtime" => $mtime]
-		);
-	}
-
-	/**
-	 * @When user :user uploads file :source to :destination with mtime :mtime using the TUS protocol on the WebDAV API
-	 *
-	 * @param string $user
-	 * @param string $source
-	 * @param string $destination
-	 * @param string $mtime Time in human readable format is taken as input which is converted into milliseconds that is used by API
-	 *
-	 * @return void
-	 */
-	public function userUploadsFileWithContentToWithMtimeUsingTUS(
-		string $user, string $source, string $destination, string $mtime
-	) {
-		$mtime = new DateTime($mtime);
-		$mtime = $mtime->format('U');
-		$user = $this->getActualUsername($user);
-		$this->userUploadsUsingTusAFileTo(
-			$user, $source, $destination, ['mtime' => $mtime]
 		);
 	}
